# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Model Computer Vision API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/batch": {
      "post": {
        "tags": [
          "Batch"
        ],
        "summary": "Submit a batch of images for background processing",
        "description": "Queue a batch of images and return a job id immediately.\n\nThe response is HTTP 202, meaning \"accepted for processing\" \u2014 not\n\"finished\". Poll ``status_url`` until ``status`` becomes ``completed`` or\n``failed``.\n\nPer-tier batch caps: free 5, basic 16, pro 32, enterprise 64.\n\nExample response::\n\n    {\n      \"job_id\": \"9f2c8a1e-...\",\n      \"status\": \"pending\",\n      \"total_items\": 20,\n      \"status_url\": \"/api/v1/batch/9f2c8a1e-...\",\n      \"estimated_seconds\": 2.4\n    }",
        "operationId": "submit_batch_api_v1_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSubmitResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Batch exceeds your tier's limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for your tier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The job queue is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batch/{job_id}": {
      "get": {
        "tags": [
          "Batch"
        ],
        "summary": "Get batch job status and results",
        "description": "Check on a batch job.\n\n``status`` progresses ``pending`` -> ``running`` -> ``completed`` or\n``failed``. While running, ``progress_percent`` updates after every image.\n\nResults appear once the job reaches a terminal state. A job can be\n``completed`` while ``failed_items`` is greater than zero: individual\nimages may fail without failing the batch, and each carries its own error.",
        "operationId": "get_batch_status_api_v1_batch__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Job id returned when the batch was submitted.",
              "title": "Job Id"
            },
            "description": "Job id returned when the batch was submitted."
          },
          {
            "name": "include_results",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include per-item results once the job has finished.",
              "default": true,
              "title": "Include Results"
            },
            "description": "Include per-item results once the job has finished."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "No job with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Batch"
        ],
        "summary": "Cancel a queued or running batch job",
        "description": "Cancel a batch job.\n\nA job that has not started is simply removed from the queue. A running job\nis terminated, so any images it had already finished are lost \u2014 the result\nis not partially retrievable after cancellation.",
        "operationId": "cancel_batch_api_v1_batch__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Job id to cancel.",
              "title": "Job Id"
            },
            "description": "Job id to cancel."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Batch Api V1 Batch  Job Id  Delete"
                }
              }
            }
          },
          "404": {
            "description": "No job with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchItem": {
        "properties": {
          "image_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Base64",
            "description": "Raw image bytes, base64-encoded. A `data:` URI prefix is accepted and stripped."
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "description": "Publicly reachable HTTPS URL to fetch the image from."
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id",
            "description": "Caller-supplied identifier echoed back in the response, useful for batches."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BatchItem",
        "description": "One entry inside a batch request."
      },
      "BatchItemResult": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index",
            "description": "Position of this item in the submitted list."
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result",
            "description": "Task-specific result payload; null when success is false."
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error code and message; null when success is true."
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "index",
          "success"
        ],
        "title": "BatchItemResult",
        "description": "Outcome for one image inside a batch.\n\nA per-item ``error`` field means one bad image does not fail the whole\nbatch \u2014 the other 63 images still return their results."
      },
      "BatchRequest": {
        "properties": {
          "model_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "pattern": "^[a-zA-Z0-9_\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name",
            "description": "Registry model name. Defaults to the task's default model."
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "pattern": "^[a-zA-Z0-9_.\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version",
            "description": "Registry version, or 'latest'. Defaults to 'latest'."
          },
          "runtime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeFormat"
              },
              {
                "type": "null"
              }
            ],
            "description": "Force a specific runtime format. Defaults to the server's preferred runtime."
          },
          "task": {
            "$ref": "#/components/schemas/TaskType",
            "description": "Which model task to run over every item."
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchItem"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items",
            "description": "Images to process. The per-request cap depends on your tier."
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "confidence_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Threshold",
            "default": 0.0
          },
          "iou_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Iou Threshold",
            "default": 0.45
          },
          "max_detections": {
            "type": "integer",
            "maximum": 300.0,
            "minimum": 1.0,
            "title": "Max Detections",
            "default": 100
          },
          "callback_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Callback Url",
            "description": "Optional HTTPS URL to POST the finished result to."
          },
          "priority": {
            "type": "integer",
            "maximum": 9.0,
            "minimum": 0.0,
            "title": "Priority",
            "description": "Queue priority; lower numbers run first.",
            "default": 5
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task",
          "items"
        ],
        "title": "BatchRequest",
        "description": "Body for ``POST /api/v1/batch``.\n\nBatches are processed by a Celery worker in the background. The endpoint\nreturns a job id immediately (HTTP 202) rather than blocking, because a\n64-image batch can take far longer than a sensible HTTP timeout."
      },
      "BatchStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "task": {
            "$ref": "#/components/schemas/TaskType"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "completed_items": {
            "type": "integer",
            "title": "Completed Items"
          },
          "failed_items": {
            "type": "integer",
            "title": "Failed Items"
          },
          "progress_percent": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Percent"
          },
          "results": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BatchItemResult"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Results",
            "description": "Populated once the job reaches a terminal state."
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Set when the whole job failed."
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "task",
          "total_items",
          "completed_items",
          "failed_items",
          "progress_percent",
          "submitted_at"
        ],
        "title": "BatchStatusResponse",
        "description": "Response for ``GET /api/v1/batch/{job_id}``."
      },
      "BatchSubmitResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "Poll this id for status and results."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Always 'pending' at submission time."
          },
          "task": {
            "$ref": "#/components/schemas/TaskType"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items",
            "description": "How many images were accepted."
          },
          "status_url": {
            "type": "string",
            "title": "Status Url",
            "description": "Ready-made URL to poll for this job."
          },
          "estimated_seconds": {
            "type": "number",
            "title": "Estimated Seconds",
            "description": "Rough completion estimate based on measured per-image latency."
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "task",
          "total_items",
          "status_url",
          "estimated_seconds",
          "correlation_id",
          "submitted_at"
        ],
        "title": "BatchSubmitResponse",
        "description": "HTTP 202 response for ``POST /api/v1/batch``.\n\nWe return a job id rather than results because batch work is queued to a\nCelery worker. Poll ``GET /api/v1/batch/{job_id}`` for progress."
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable code, e.g. 'IMAGE_TOO_LARGE'."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Friendly, user-facing explanation."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Inner object of the standard error envelope."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The single error shape every failing endpoint returns."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "description": "Lifecycle of an asynchronous batch job."
      },
...
      "RuntimeFormat": {
        "type": "string",
        "enum": [
          "torch",
          "torch_int8",
          "onnx",
          "onnx_int8",
          "tensorrt"
        ],
        "title": "RuntimeFormat",
        "description": "Which compiled form of a model is executing."
      },
...
      "TaskType": {
        "type": "string",
        "enum": [
          "classification",
          "detection",
          "similarity"
        ],
        "title": "TaskType",
        "description": "The three computer-vision tasks this system serves."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key issued to your account. Determines your tier."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived JWT access token."
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
}